Refactoring
reimplemented tests
reimplemented generate quadgrams funktion
